Added GuestController, GuestRepository and minor fixes.

+++ src/main/java/com/molvenolakeresort/repositories/hotel/GuestRepository.java
@@ -3,5 +3,6 @@ package com.molvenolakeresort.repositories.hotel;
 import com.molvenolakeresort.models.hotel.Guest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GuestRepository extends JpaRepository<Guest, Long> {
+public interface GuestRepository extends JpaRepository<Guest, Integer> {
+
 }
